Chore: move PrometheusListView specific utils in the correct directory
